Moves chess files, initial Tetris board, movement, and collision logic
