+++ src/chess/classes/pieces.js
