+++ src/chess/classes/traverseFunctions.js
