+++ src/css/tetris.css
@@ -1,0 +1,17 @@
+.Square{
+  display: flex;
+  flex-direction: column;
+  height: 35px;
+  width: 35px;
+  background-color: black;
+  margin: 2px;
+}
+
+.Row{
+  display: flex;
+  flex-direction: row;
+}
+
+.Purple{
+  background-color: purple;
+}
